+++ MSP App/API/Controllers/DeviceController.cs
@@ -1,10 +1,13 @@
-﻿using API.Services;
+﻿using API.Attributes.HasPermission;
+using API.Services;
 using API.Services.Interfaces;
+using Azure.Core;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using SharedComponents.DbServices;
 using SharedComponents.Entities;
 using SharedComponents.ResponseEntities;
+using SharedComponents.Results;
 using SharedComponents.WebEntities.Requests.DeviceRequests;
 
 namespace API.Controllers
@@ -18,21 +21,31 @@ namespace API.Controllers
         private readonly IDbTenantService _dbTenantService;
         private readonly IDbInstallationKeyService _dbInstallationKeyService;
         private readonly IHTTPDeviceService _httpDeviceService;
+        private readonly IAuthService _authService;
 
         public DeviceController(IDbDeviceService dbDeviceService, IDbTenantService dbTenantService,
-            IDbInstallationKeyService dbInstallationKeyService, IHTTPDeviceService httpDeviceService)
+            IDbInstallationKeyService dbInstallationKeyService, IHTTPDeviceService httpDeviceService,
+            IAuthService authService)
         {
             _dbDeviceService = dbDeviceService;
             _dbTenantService = dbTenantService;
             _dbInstallationKeyService = dbInstallationKeyService;
             _httpDeviceService = httpDeviceService;
+            _authService = authService;
         }
 
         [HttpPost("")]
+        [HasPermission("Device.Create.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> CreateAsync([FromBody] DeviceCreateRequest request)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), request.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 Tenant? tenant = await _dbTenantService.GetAsync(request.TenantId);
                 if (tenant == null)
                 {
@@ -59,7 +72,7 @@ namespace API.Controllers
                     return Unauthorized("Inactive Installation Key.");
                 }
 
-                if(installationKey.Type != InstallationKeyType.Server || installationKey.Type != InstallationKeyType.Device)
+                if(installationKey.Type != InstallationKeyType.Server && installationKey.Type != InstallationKeyType.Device)
                 {
                     return Unauthorized("Invalid Installation Key Type.");
                 }
@@ -137,6 +150,7 @@ namespace API.Controllers
         }
 
         [HttpPut("")]
+        [HasPermission("Device.Update.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> UpdateAsync([FromBody] DeviceUpdateRequest request)
         {
             if (ModelState.IsValid)
@@ -146,6 +160,12 @@ namespace API.Controllers
                 {
                     return NotFound("Device not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (device.DeviceInfo == null)
                 {
                     return NotFound("DeviceInfo not found.");
@@ -158,7 +178,10 @@ namespace API.Controllers
                         Device? server = devices.FirstOrDefault(d => d.DeviceInfo!.Type == DeviceType.Server);
                         if (server != null)
                         {
-                            return BadRequest("Server already registered.");
+                            if(server.Id != request.Id)
+                            {
+                                return BadRequest("Server already registered.");
+                            }
                         }
                     }
                 }
@@ -201,6 +224,7 @@ namespace API.Controllers
         }
 
         [HttpPut("Status")]
+        [HasPermission("Device.Update-Status.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> UpdateStatusAsync([FromBody] DeviceUpdateStatusRequest request)
         {
             if (ModelState.IsValid)
@@ -210,6 +234,12 @@ namespace API.Controllers
                 {
                     return NotFound("Device not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (device.DeviceInfo == null)
                 {
                     return NotFound("DeviceInfo not found.");
@@ -226,10 +256,22 @@ namespace API.Controllers
         }
 
         [HttpDelete("{id}")]
+        [HasPermission("Device.Delete.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             if (ModelState.IsValid)
             {
+                Device? device = await _dbDeviceService.GetAsync(id);
+                if (device == null)
+                {
+                    return NotFound("Device not found.");
+                }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (await _dbDeviceService.DeleteAsync(id))
                 {
                     return Ok($"Device deleted successfully.");
@@ -240,6 +282,7 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}")]
+        [HasPermission("Device.Get.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAsync(int id)
         {
             if (ModelState.IsValid)
@@ -247,6 +290,12 @@ namespace API.Controllers
                 Device? device = await _dbDeviceService.GetAsync(id);
                 if (device != null)
                 {
+                    // Authentication check using roles + permissions
+                    if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                    {
+                        return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                    }
+                    // --- End of authentication check ---
                     return Ok(device);
                 }
                 return NotFound("Device not found.");
@@ -255,14 +304,34 @@ namespace API.Controllers
         }
 
         [HttpGet("")]
+        [HasPermission("Device.Get-All.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllAsync()
         {
             if (ModelState.IsValid)
             {
-                ICollection<Device>? devices = await _dbDeviceService.GetAllAsync();
+                var tenantIds = await _authService.GetUserAccessibleTenantsAsync(Request.Headers["Authorization"].ToString());
+                if (tenantIds == null)
+                {
+                    return new CustomForbidResult("User does not have any tenant permissions");
+                }
+                List<Device> devices = new List<Device>();
+                foreach (var tenantId in tenantIds)
+                {
+                    if (tenantId != null)
+                    {
+                        var tenantDevices = await _dbDeviceService.GetAllByTenantIdAsync((int)tenantId);
+                        if (tenantDevices != null)
+                        {
+                            devices.AddRange(tenantDevices);
+                        }
+                    }
+                }
                 if (devices != null)
                 {
-                    return Ok(devices);
+                    if (devices.Any())
+                    {
+                        return Ok(devices.Distinct());
+                    }
                 }
                 return NotFound("No devices found.");
             }
@@ -270,10 +339,17 @@ namespace API.Controllers
         }
 
         [HttpGet("By-Tenant/{tenantId}")]
+        [HasPermission("Device.Get-By-Tenant.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllByTenantIdAsync(int tenantId)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), tenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 ICollection<Device>? devices = await _dbDeviceService.GetAllByTenantIdAsync(tenantId);
                 if (devices != null)
                 {
@@ -288,6 +364,7 @@ namespace API.Controllers
         }
 
         [HttpPost("{id}/Refresh")]
+        [HasPermission("Device.Refresh.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> RefreshAsync(int id)
         {
             if (ModelState.IsValid)
@@ -297,6 +374,12 @@ namespace API.Controllers
                 {
                     return NotFound("Device not found.");
                 }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<DeviceController>(Request.Headers["Authorization"].ToString(), device.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (device.DeviceInfo == null)
                 {
                     return NotFound("DeviceInfo not found.");
