+++ MSP App/API/Controllers/InstallationKeyController.cs
@@ -1,9 +1,12 @@
+﻿using API.Attributes.HasPermission;
 ﻿using API.Services;
+using API.Services.Interfaces;
 using Azure.Core;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using SharedComponents.DbServices;
 using SharedComponents.Entities;
+using SharedComponents.Results;
 using SharedComponents.WebEntities.Requests.InstallationKeyRequests;
 
 namespace API.Controllers
@@ -15,14 +18,17 @@ namespace API.Controllers
     {
         private readonly IDbInstallationKeyService _dbInstallationKeyService;
         private readonly IDbTenantService _dbTenantService;
+        private readonly IAuthService _authService;
 
-        public InstallationKeyController(IDbInstallationKeyService dbInstallationKeyService, IDbTenantService dbTenantService)
+        public InstallationKeyController(IDbInstallationKeyService dbInstallationKeyService, IDbTenantService dbTenantService, IAuthService authService)
         {
             _dbInstallationKeyService = dbInstallationKeyService;
             _dbTenantService = dbTenantService;
+            _authService = authService;
         }
 
         [HttpPost("")]
+        [HasPermission("InstallationKey.Create.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> CreateAsync([FromBody] InstallationKeyCreateRequest request)
         {
             if( ModelState.IsValid)
@@ -32,7 +38,13 @@ namespace API.Controllers
                 {
                     return NotFound("Tenant not found.");
                 }
-                if(request.Type == null)
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<InstallationKeyController>(Request.Headers["Authorization"].ToString(), tenant.Id))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
+                if (request.Type == null)
                 {
                     return BadRequest("InstallationKey Type is required.");
                 }
@@ -62,6 +74,7 @@ namespace API.Controllers
         }
 
         [HttpPut("")]
+        [HasPermission("InstallationKey.Update.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> UpdateAsync([FromBody] InstallationKeyUpdateRequest request)
         {
             if (ModelState.IsValid)
@@ -71,7 +84,25 @@ namespace API.Controllers
                 {
                     return NotFound("Installation key not found.");
                 }
-                if(installationKey.IsDeleted)
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<InstallationKeyController>(Request.Headers["Authorization"].ToString(), installationKey.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
+                
+                if (request.Type == null)
+                {
+                    return BadRequest("InstallationKey Type is required.");
+                }
+                if (request.Type.HasValue)
+                {
+                    if (!Enum.IsDefined(typeof(InstallationKeyType), request.Type.Value))
+                    {
+                        return BadRequest("Invalid InstallationKey Type.");
+                    }
+                }
+                if (installationKey.IsDeleted)
                 {
                     return BadRequest("Installation key is deleted.");
                 }
@@ -87,10 +118,22 @@ namespace API.Controllers
         }
 
         [HttpDelete("{id}")]
+        [HasPermission("InstallationKey.Delete.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> DeleteAsync(int id)
         {
             if (ModelState.IsValid)
             {
+                InstallationKey? installationKey = await _dbInstallationKeyService.GetAsync(id);
+                if (installationKey == null)
+                {
+                    return NotFound("Installation key not found.");
+                }
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<InstallationKeyController>(Request.Headers["Authorization"].ToString(), installationKey.TenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 if (await _dbInstallationKeyService.DeleteAsync(id))
                 {
                     return Ok($"Installation key deleted successfully.");
@@ -101,6 +144,7 @@ namespace API.Controllers
         }
 
         [HttpGet("{id}")]
+        [HasPermission("InstallationKey.Get.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAsync(int id)
         {
             if (ModelState.IsValid)
@@ -108,6 +152,12 @@ namespace API.Controllers
                 InstallationKey? installationKey = await _dbInstallationKeyService.GetAsync(id);
                 if (installationKey != null)
                 {
+                    // Authentication check using roles + permissions
+                    if (!await _authService.UserHasPermissionAsync<InstallationKeyController>(Request.Headers["Authorization"].ToString(), installationKey.TenantId))
+                    {
+                        return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                    }
+                    // --- End of authentication check ---
                     return Ok(installationKey);
                 }
                 return NotFound("Installation key not found.");
@@ -116,16 +166,33 @@ namespace API.Controllers
         }
 
         [HttpGet("")]
+        [HasPermission("InstallationKey.Get-All.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllAsync()
         {
             if (ModelState.IsValid)
             {
-                ICollection<InstallationKey>? installationKeys = await _dbInstallationKeyService.GetAllAsync();
+                var tenantIds = await _authService.GetUserAccessibleTenantsAsync(Request.Headers["Authorization"].ToString());
+                if (tenantIds == null)
+                {
+                    return new CustomForbidResult("User does not have any tenant permissions");
+                }
+                List<InstallationKey> installationKeys = new List<InstallationKey>();
+                foreach (var tenantId in tenantIds)
+                {
+                    if (tenantId != null)
+                    {
+                        var tenantInstallationKeys = await _dbInstallationKeyService.GetAllByTenantIdAsync((int)tenantId);
+                        if (tenantInstallationKeys != null)
+                        {
+                            installationKeys.AddRange(tenantInstallationKeys);
+                        }
+                    }
+                }
                 if (installationKeys != null)
                 {
                     if (installationKeys.Any())
                     {
-                        return Ok(installationKeys);
+                        return Ok(installationKeys.Distinct());
                     }
                 }
                 return NotFound("Installation keys not found.");
@@ -134,10 +201,17 @@ namespace API.Controllers
         }
 
         [HttpGet("By-Tenant/{tenantId}")]
+        [HasPermission("InstallationKey.Get-By-Tenant.Permission", PermissionType.Tenant)]
         public async Task<IActionResult> GetAllByTenantIdAsync(int tenantId)
         {
             if (ModelState.IsValid)
             {
+                // Authentication check using roles + permissions
+                if (!await _authService.UserHasPermissionAsync<InstallationKeyController>(Request.Headers["Authorization"].ToString(), tenantId))
+                {
+                    return new CustomForbidResult("User do not have access to this feature for the specified tenant");
+                }
+                // --- End of authentication check ---
                 Tenant? tenant = await _dbTenantService.GetAsync(tenantId);
                 if (tenant == null)
                 {
